Add cohort members retrival feature
